# Use "Faster Subgroup Checks for BLS12-381" for `is_torsion_free`