This uses @ebfull's method for [faster subgroup checks](https://eprint.iacr.org/2019/814)

I didn't know if maybe there's a specific reason you all aren't already using these methods here. If so, sorry for the noise!

On my Core i7 Macbook Pro, the "G1Affine subgroup check" benchmark takes:
- 390us on `main`
  - Note, though, that the changes introduced in #79 make this 211us
- 90us with these changes

Similarly, the "G2Affine subgroup check" benchmark takes:
- 1.6ms on `main`
- 175us with these changes